refactor(tests): extract workspace_scan.rs (16 tests + v6c probe)
Move collect_documents_* suite and v6c_speedup_probe from mod.rs into
engine/tests/workspace_scan.rs. Module header carries the v6c-specific
hazard: do NOT env_mutation inside these tests (cargo parallel test,
process-global env race).

Tests moved (all now at engine::tests::workspace_scan::*):
  collect_documents_includes_json_alongside_markdown
  collect_documents_markdown_unchanged
  collect_documents_hard_ignore_baseline
  collect_documents_respects_gitignore
  collect_documents_baseline_without_ignore_file
  collect_documents_output_is_sorted
  collect_documents_terminates_on_symlink_cycle
  collect_documents_empty_dir
  collect_documents_nonexistent_root_does_not_panic
  collect_documents_root_is_a_file
  collect_documents_unicode_filenames
  collect_documents_is_deterministic_across_runs
  collect_documents_deep_nesting
  collect_documents_broken_symlink_is_skipped
  collect_documents_file_named_target_is_not_filtered
  v6c_speedup_probe (still #[ignore]d)

Tests: 135 passed / 0 failed / 1 ignored (baseline unchanged).

marky-n1h step 5.
